08_DynamicPrompt/pyproject.toml
[project]
name = "08-dynamicprompt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.3",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
]
